Scripts内の名前空間を整理(フォルダと合わせる)
