+++ Assets/New Folder/Scripts/Extends/Actors/Platformers/ActorActionStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6a069562cfb211147aff76dca1c9ab2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
