+++ Assets/New Folder/Scripts/Extends/Actors/Platformers/ButtonOperation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e03c9dabca20564d83dc7cdb603d919
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
