+++ Assets/New Folder/Scripts/Extends/Actors/Platformers/CurrentStateManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fcbfa9540aa1a4b4487b88e3aa32f5fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
