+++ Assets/New Folder/Scripts/TestScripts/Test.cs
@@ -1,0 +1,49 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+namespace TestScripts
+{
+    public class Test : MonoBehaviour
+    {
+
+        public AudioClip sound1;
+        public AudioClip sound2;
+        public AudioClip sound3;
+        public AudioClip sound4;
+
+        AudioSource audioSource;
+
+        void Start()
+        {
+            //Componentを取得
+            audioSource = GetComponent<AudioSource>();
+        }
+
+        void Update()
+        {
+            // 左
+            if (Input.GetKey(KeyCode.LeftArrow))
+            {
+                audioSource.PlayOneShot(sound1);
+            }
+            // 右
+            if (Input.GetKey(KeyCode.RightArrow))
+            {
+                audioSource.PlayOneShot(sound2);
+            }
+            // 上
+            if (Input.GetKey(KeyCode.UpArrow))
+            {
+                audioSource.PlayOneShot(sound3);
+            }
+            // 下
+            if (Input.GetKey(KeyCode.DownArrow))
+            {
+                audioSource.PlayOneShot(sound4);
+            }
+        }
+
+    }
+}
