+++ Assets/New Folder/Scripts/TestScripts/Test_CountCoins.cs.meta
