+++ Assets/New Folder/Scripts/TestScripts/Test_GameSave.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Game;
+
+
+namespace TestScripts
+{
+    public class Test_GameSave : MonoBehaviour
+    {
+        public void Save()
+        {
+            GameController.SaveGame();
+        }
+    }
+}
