--- Assets/New Folder/Scripts/_Actor/IPlayer.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IPlayer{
-
-    /// <summary>
-    /// プレイヤーの死亡
-    /// </summary>
-    void Death();
-
-    /// <summary>
-    /// 死亡時に生成するエフェクトオブジェクト
-    /// </summary>
-    GameObject DeathEffectObject { get;}
-
-    /// <summary>
-    /// 自身は死んでいるかどうか
-    /// </summary>
-    bool IsDeath { get; set; }
-}
