--- Assets/New Folder/Scripts/_TestScripts/Test_GameSave.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using Game;
-
-
-public class Test_GameSave : MonoBehaviour
-{
-    public void Save()
-    {
-        GameController.SaveGame();
-    }
-}
